stocksense/requirements.txt
flask
yfinance
launchdarkly-server-sdk
python-dotenv
pandas
